Move the implementation of IVsTextViewFilter.GetDataTipText
This implementation lived in VsLanguageDebugInfo, but wasn't dependent
on anything in that, really. Moving this and inlining the code means
we can remove the need for an AbstractTextViewFilter to hold onto
an AbstractLanguageService which is a problem.
