+++ src/VisualStudio/Core/Def/Implementation/AbstractVsTextViewFilter`2.cs
@@ -4,10 +4,17 @@ using System;
 using System.Threading;
 using Microsoft.CodeAnalysis.Editor;
 using Microsoft.CodeAnalysis.Editor.Host;
+using Microsoft.CodeAnalysis.Editor.Implementation.Debugging;
 using Microsoft.CodeAnalysis.Editor.Shared.Extensions;
 using Microsoft.CodeAnalysis.ErrorReporting;
+using Microsoft.CodeAnalysis.Internal.Log;
+using Microsoft.CodeAnalysis.Text;
+using Microsoft.CodeAnalysis.Text.Shared.Extensions;
 using Microsoft.VisualStudio.ComponentModelHost;
+using Microsoft.VisualStudio.LanguageServices.Implementation.Extensions;
 using Microsoft.VisualStudio.LanguageServices.Implementation.LanguageService;
+using Microsoft.VisualStudio.Shell;
+using Microsoft.VisualStudio.Shell.Interop;
 using Microsoft.VisualStudio.Text;
 using Microsoft.VisualStudio.Text.Editor;
 using Microsoft.VisualStudio.TextManager.Interop;
@@ -41,14 +48,7 @@ namespace Microsoft.VisualStudio.LanguageServices.Implementation
                     return VSConstants.E_INVALIDARG;
                 }
 
-                var debugInfo = LanguageService.LanguageDebugInfo;
-                if (debugInfo == null)
-                {
-                    pbstrText = null;
-                    return VSConstants.E_FAIL;
-                }
-
-                return GetDataTipTextImpl(pSpan, debugInfo, out pbstrText);
+                return GetDataTipTextImpl(pSpan, out pbstrText);
             }
             catch (Exception e) when (FatalError.ReportWithoutCrash(e) && false)
             {
@@ -56,7 +56,7 @@ namespace Microsoft.VisualStudio.LanguageServices.Implementation
             }
         }
 
-        protected virtual int GetDataTipTextImpl(TextSpan[] pSpan, AbstractLanguageService<TPackage, TLanguageService>.VsLanguageDebugInfo debugInfo, out string pbstrText)
+        protected virtual int GetDataTipTextImpl(TextSpan[] pSpan, out string pbstrText)
         {
             var subjectBuffer = WpfTextView.GetBufferContainingCaret();
             if (subjectBuffer == null)
@@ -65,10 +65,10 @@ namespace Microsoft.VisualStudio.LanguageServices.Implementation
                 return VSConstants.E_FAIL;
             }
 
-            return GetDataTipTextImpl(subjectBuffer, pSpan, debugInfo, out pbstrText);
+            return GetDataTipTextImpl(subjectBuffer, pSpan, out pbstrText);
         }
 
-        protected int GetDataTipTextImpl(ITextBuffer subjectBuffer, TextSpan[] pSpan, AbstractLanguageService<TPackage, TLanguageService>.VsLanguageDebugInfo debugInfo, out string pbstrText)
+        protected int GetDataTipTextImpl(ITextBuffer subjectBuffer, TextSpan[] pSpan, out string pbstrText)
         {
             pbstrText = null;
 
@@ -80,7 +80,63 @@ namespace Microsoft.VisualStudio.LanguageServices.Implementation
                 return VSConstants.E_FAIL;
             }
 
-            return debugInfo.GetDataTipText(vsBuffer, pSpan, out pbstrText);
+            using (Logger.LogBlock(FunctionId.Debugging_VsLanguageDebugInfo_GetDataTipText, CancellationToken.None))
+            {
+                pbstrText = null;
+                if (pSpan == null || pSpan.Length != 1)
+                {
+                    return VSConstants.E_INVALIDARG;
+                }
+
+                var result = VSConstants.E_FAIL;
+                string pbstrTextInternal = null;
+
+                var waitIndicator = ComponentModel.GetService<IWaitIndicator>();
+
+                waitIndicator.Wait(
+                    title: ServicesVSResources.Debugger,
+                    message: ServicesVSResources.Getting_DataTip_text,
+                    allowCancel: true,
+                    action: waitContext =>
+                {
+                    IServiceProvider serviceProvider = ComponentModel.GetService<SVsServiceProvider>();
+                    var debugger = (IVsDebugger)serviceProvider.GetService(typeof(SVsShellDebugger));
+                    var debugMode = new DBGMODE[1];
+
+                    var cancellationToken = waitContext.CancellationToken;
+                    if (ErrorHandler.Succeeded(debugger.GetMode(debugMode)) && debugMode[0] != DBGMODE.DBGMODE_Design)
+                    {
+                        var textSpan = pSpan[0];
+
+                        var textSnapshot = subjectBuffer.CurrentSnapshot;
+                        var document = textSnapshot.GetOpenDocumentInCurrentContextWithChanges();
+
+                        if (document != null)
+                        {
+                            var languageDebugInfo = document.Project.LanguageServices.GetService<ILanguageDebugInfoService>();
+                            if (languageDebugInfo != null)
+                            {
+                                var spanOpt = textSnapshot.TryGetSpan(textSpan);
+                                if (spanOpt.HasValue)
+                                {
+                                    var dataTipInfo = languageDebugInfo.GetDataTipInfoAsync(document, spanOpt.Value.Start, cancellationToken).WaitAndGetResult(cancellationToken);
+                                    if (!dataTipInfo.IsDefault)
+                                    {
+                                        var resultSpan = dataTipInfo.Span.ToSnapshotSpan(textSnapshot);
+                                        var textOpt = dataTipInfo.Text;
+
+                                        pSpan[0] = resultSpan.ToVsTextSpan();
+                                        result = debugger.GetDataTipValue((IVsTextLines)vsBuffer, pSpan, textOpt, out pbstrTextInternal);
+                                    }
+                                }
+                            }
+                        }
+                    }
+                });
+
+                pbstrText = pbstrTextInternal;
+                return result;
+            }
         }
 
         int IVsTextViewFilter.GetPairExtents(int iLine, int iIndex, TextSpan[] pSpan)
