+++ src/VisualStudio/Core/Def/Implementation/LanguageService/AbstractLanguageService`2.VsLanguageDebugInfo.cs
@@ -13,12 +13,10 @@ using Microsoft.CodeAnalysis.Text.Shared.Extensions;
 using Microsoft.VisualStudio.LanguageServices.Implementation.Debugging;
 using Microsoft.VisualStudio.LanguageServices.Implementation.Extensions;
 using Microsoft.VisualStudio.LanguageServices.Implementation.Utilities;
-using Microsoft.VisualStudio.Shell.Interop;
 using Roslyn.Utilities;
 using IVsDebugName = Microsoft.VisualStudio.TextManager.Interop.IVsDebugName;
 using IVsEnumBSTR = Microsoft.VisualStudio.TextManager.Interop.IVsEnumBSTR;
 using IVsTextBuffer = Microsoft.VisualStudio.TextManager.Interop.IVsTextBuffer;
-using IVsTextLines = Microsoft.VisualStudio.TextManager.Interop.IVsTextLines;
 using RESOLVENAMEFLAGS = Microsoft.VisualStudio.TextManager.Interop.RESOLVENAMEFLAGS;
 using VsTextSpan = Microsoft.VisualStudio.TextManager.Interop.TextSpan;
 
@@ -371,63 +369,6 @@ namespace Microsoft.VisualStudio.LanguageServices.Implementation.LanguageService
                 }
 
                 return VSConstants.E_NOTIMPL;
-            }
-
-            public int GetDataTipText(IVsTextBuffer pBuffer, VsTextSpan[] pSpan, out string pbstrText)
-            {
-                using (Logger.LogBlock(FunctionId.Debugging_VsLanguageDebugInfo_GetDataTipText, CancellationToken.None))
-                {
-                    pbstrText = null;
-                    if (pSpan == null || pSpan.Length != 1)
-                    {
-                        return VSConstants.E_INVALIDARG;
-                    }
-
-                    var result = VSConstants.E_FAIL;
-                    string pbstrTextInternal = null;
-
-                    _waitIndicator.Wait(
-                        title: ServicesVSResources.Debugger,
-                        message: ServicesVSResources.Getting_DataTip_text,
-                        allowCancel: true,
-                        action: waitContext =>
-                    {
-                        var debugger = _languageService.Debugger;
-                        var debugMode = new DBGMODE[1];
-
-                        var cancellationToken = waitContext.CancellationToken;
-                        if (ErrorHandler.Succeeded(debugger.GetMode(debugMode)) && debugMode[0] != DBGMODE.DBGMODE_Design)
-                        {
-                            var editorAdapters = _languageService.EditorAdaptersFactoryService;
-
-                            var textSpan = pSpan[0];
-                            var subjectBuffer = editorAdapters.GetDataBuffer(pBuffer);
-
-                            var textSnapshot = subjectBuffer.CurrentSnapshot;
-                            var document = textSnapshot.GetOpenDocumentInCurrentContextWithChanges();
-
-                            if (document != null)
-                            {
-                                var spanOpt = textSnapshot.TryGetSpan(textSpan);
-                                if (spanOpt.HasValue)
-                                {
-                                    var dataTipInfo = _languageDebugInfo.GetDataTipInfoAsync(document, spanOpt.Value.Start, cancellationToken).WaitAndGetResult(cancellationToken);
-                                    if (!dataTipInfo.IsDefault)
-                                    {
-                                        var resultSpan = dataTipInfo.Span.ToSnapshotSpan(textSnapshot);
-                                        var textOpt = dataTipInfo.Text;
-
-                                        pSpan[0] = resultSpan.ToVsTextSpan();
-                                        result = debugger.GetDataTipValue((IVsTextLines)pBuffer, pSpan, textOpt, out pbstrTextInternal);
-                                    }
-                                }
-                            }
-                        }
-                    });
-
-                    pbstrText = pbstrTextInternal;
-                    return result;
-                }
             }
         }
     }
